refactor(run-once): harden recovery repair orchestration
